+++ app/controllers/concerns/authentication_and_sso_concerns.rb
@@ -4,37 +4,14 @@
 # to this responsibility alone.
 module AuthenticationAndSSOConcerns
   extend ActiveSupport::Concern
-  include ActionController::RequestForgeryProtection
   include ActionController::Cookies
 
   included do
-    before_action :validate_csrf_token!, if: lambda {
-      ActionController::Base.allow_forgery_protection && request.method != 'GET'
-    }
     before_action :authenticate
     before_action :set_session_expiration_header
-    after_action :set_csrf_cookie
   end
 
   protected
-
-  def set_csrf_cookie
-    cookies['X-CSRF-Token'] = {
-      value: form_authenticity_token,
-      expires: 1.day.from_now,
-      secure: true
-    }
-  end
-
-  def validate_csrf_token!
-    if request.headers['X-CSRF-Token'].nil? || request.headers['X-CSRF-Token'] != cookies['X-CSRF-Token']
-      # raise ActionController::InvalidAuthenticityToken
-
-      # for now we are just logging when there's no CSRF protection
-      # when this is going to be enforced return a meaningful error (and turn up logging level)
-      log_message_to_sentry('Request susceptible to CSRF', :info, controller: self.class, action: action_name)
-    end
-  end
 
   def authenticate
     validate_session || render_unauthorized
